Credits (#51)
Add a CREDITS file, and render the credits after the exposition on
victory.

Anything you add to the CREDITS file will display verbatim in the
credits, except a rendering bug makes the spaces render at a different
spacing than the rest of the font. Apparently this is a known MonoGame
issue.